Added user database and events database. Added Login, List and Edit views in EventsReact (they are still mostly from the example).
Managed to send the current username to EventList, EventEdit and EventView.
While getting the list of ALL events the server is now aware of the username that makes the request (it will set the field canEdit on true for that user).
Other routes will know the user making the request in the same way (by decoding the username from token).